Skip tab-completion tests if envar SKIP_READLINE_TESTS is defined.
Experience so far suggests that getting these tests to pass on
all libedit versions that are out there may be impossible, or
require dumbing down the tests to the point of uselessness.
So we need to provide a way to skip them when the user knows they'll
fail.  An environment variable is probably the most convenient way
to deal with this; it's easy for, e.g., a buildfarm animal's
configuration to set up.

Discussion: https://postgr.es/m/9594.1578586797@sss.pgh.pa.us

(cherry-picked from e7ee433137b6eb3678deeebc68be53d58aa1e5d3)

Ticket: GG-561