Fix pergroupstate for non-strict transition functions
On some datasets GP fails with segfault on serialization of a
HHashAgg table entries to spillfiles. This problem is caused by a
pergroupstate inconsistency in some entries: transValue is 0 when
transValueIsNull and noTransValue are false.

The reason of this inconsistency is a wrong fcinfo->isnull
status returned by some of non-strict *_combine transition
functions. The first part of the problem is that *_combine functions
do not modify fcinfo->isnull state but can return NULL. The second
part of it is that fcinfo->isnull in invoke_agg_trans_func() is always
set false by default for non-strict transition functions. As a result
pergroupstate can become inconsistent and cause segfailts.

We can get information whether a transition function returns NULL
or not only from its internals. So this commit fixes *_combine
functions to set fcinfo->isnull correctly.

Co-authored-by: Dmitry Ursegov <uda@arenadata.io>