Limit gxact number on master with MaxBackends.
Previously we assign it as max_prepared_xacts. It is used to initialize some
2pc related shared memory. For example the array shmCommittedGxactArray is
created with this length and that array is used to collect not-yet "forgotten"
distributed transactions during master/standby recovery, but the array length
might be problematic since:

1. master max_prepared_xacts might be different than segment max_prepared_xacts
(although the document does not suggest it there is no code to enforce that),
so the array on master might not be able to accommodate the distributed
transactions.

2. Even master max_prepared_xacts is equal to segment max_prepared_xacts as
usual.  It is possible some distributed transactions use just partial gang so
the total distributed transactions might be larger (and even much larger) than
max_prepared_xacts. The document says max_prepared_xacts should be greater than
max_connections but there is no code to enforce that.

To fix that we use MaxBackends for the gxact number on master. We may just use
guc max_connections (MaxBackends includes number for autovacuum workers and bg
workers additionally besides guc max_connections), but I'm conservatively using
MaxBackends.  since this issue is annoying - standby can not recover even after
reboot unless we temporarily increase the guc max_prepared_transactions value.

Backport to 6X of 761c889e242aa287ab1a2dfdf7f111c4569768fd

Co-authored-by: Denis Smirnov <sd@arenadata.io>