pg_partitions_nolock: pg_partitions without locks on relations
Implement a different version of 'pg_get_expr' which does not process
'Var' nodes, but also does not require relation OID. Its purpose is to
deparse expressions when called by 'pg_partitions' VIEW which do not
contain 'Var' nodes.

The method is packed into an extension, which renames the original GPDB
view 'pg_partitions' to 'pg_partitions_lock' and recreates
'pg_partitions', attaching the non-locking version of 'pg_get_expr' to
it.