Fix bug related to incorrect use of WaitLatchOrSocket function
This commit is fix for #144 PR. Missed a problem with incorrect use of the function.
We always have to pass the WL_SOCKET_READABLE flag to WaitLatchOrSocket function when
we waiting on sockets. The reason for this is that EOF and error conditions are reported
only via WL_SOCKET_READABLE. This patch fixes that problem.