pg_rewind: avoid removing "pg_log" files
"pg_log" is part of excluded directory list. The semantics of exclude
file/directory list is to avoid copy from source to target. But also
as side effect if those files are present in target, they are removed
as well.

This semantic is good for most of things but for "pg_log" directory
don't wish to get them removed. Hence, adding logic to add this
exception.

Backport 6f29a6f to 6X.

Backported-by: Maksim Milyutin <mma@arenadata.io>