Demonstrate how catalog snapshot corrupts AO/AOCO data
In 5X a snapshot satisfies "now" was used when transaction extracted
AO/AOCO segment file eof. In other words "now" allowed us to see all
committed transactions. When this code was ported to 6X a catalog
snapshot was chosen for this purpose. As a result we don't see committed
transactions that started after the snapshot has been taken. As a
result a transaction can change segment file eof between the moment
a snapshot was taken and an exclusive file segment lock was hold.
So AO/AOCO in 6X suffers from data corruption in concurrent DML
queries.

This is the first commit demonstrating the problem (all added tests
should fail).