Detect client disconnection while running query and immediately interrupt its execution (#198)
Fixes the case when the server is executing a lengthy query and the
client breaks the connection. The operating system will be aware that
the connection is no more, but postgres node doesn't notice this,
because it doesn't try to read from or write to the socket while running
query. So we'll get a zombie connection. In theory, the query could be
one that runs for a million years, continues to chew up CPU and I/O and
occupies a connection slot that's sad. Worse still, a sent query might
be modifiable and not return any data, then it might be surprising for
disconnected client that his previously sent modification will be
accepted at some point later - at completion of execution. For these
reasons, the query have to be interrupted as much earlier as possible.

The patch provides a new GUC check_client_connection_interval that can
be used to periodically check via CLIENT_CHECK_CONNECTION_TIMEOUT
interrupts whether the client connection has gone away, while running
very long queries. It is disabled by default.
For non-locking check of socket state the patch uses a non-standard
Linux extension (also adopted by at least one other OS) - POLLRDHUP
option that is not defined by POSIX.

Backport from PostgreSQL commits:
 - https://git.postgresql.org/gitweb/?p=postgresql.git;a=commit;h=c30f54ad732ca5c8762bb68bbe0f51de9137dd72
 - https://git.postgresql.org/gitweb/?p=postgresql.git;a=commit;h=22f6f2c1ccb56e0d6a159d4562418587e4b10e01