use system gcc 4.8.5 for tests
https://github.com/arenadata/gpdb/commit/1e375f08bfe2509cf70386aea794cf5ed73591a7 removes gcc sourcing from test pipelines
Upstream are going to use docker environment variables to inject PATH and LD_LIBRARY_PATH
But this variables is applied only for root, but not for gpadmin user that run test too.
But gpadmin uses gcc to build debug extensions during test run.
As a result, gpadmin uses system gcc with pg_config flags from GCC 7