Add check to ensure stats are not null in implementation stage of Orca (#12491)
In some edge cases in Orca, groups may not have had statistics derived
during the exploration stage. This is ok in many cases, as these groups
may not be used during implementation or their stats may not be accessed
during implementation. However, there are some cases where we do need
these stats, and trying to access them resulted in a crash.

This commit adds a defensive check to fall back to planner if we
encounter such a scenario, similar to checks added in
a6f92be1b8327aadc7l.

Co-authored-by: Chris Hajas <chajas@vmware.com>
Co-authored-by: Shreedhar Hardikar <shardikar@vmware.com>