Fix memory leak for DynamicBitmapHeapScan and DynamicSeqScan. (#12562)
Commit db5163474d did a big refactor of DynamicXXXScan, unfortunately,
it introduces memory leak during Executor running stage for
ExecDynamicBitmapHeapScan and ExecDynamicSeqScan (ExecDynamicIndex does
not leak memory). Use ExecDynamicSeqScan to demonstrate the issue, it
will init a new seqscan state for each new partition (some memroy will
be allocated during this operation), this is not a good pattern as I
can see. Postgres only invoke ExecInitXXX during InitPlan(), never in
ExectorRun()'s lifetime. The point is ExecDynamicSeqScan might appear
as the inner plan of Nestloop join which means it might be rescaned
many many times. So we cannot init a new seqscan state each time for
a new partition or we have to clean it up totally after a single partition
is finished one time.

We have tried to use the same skill in ExecDynamicIndex to fix this, the
skill is to wrap the init procedure in a memory context. Things are not
that easy for DynamicSeqScan adn DynamicBitmapHeapScan, we spend some time
on the memory context fix method but not manage to make it correct.

Finally we decide to use a cache method to fix this issue. Also take
DynamicSeqScan as an example to state our method. We build a hashtable
during ExecInitDynamicSeqScan and when init new seqscan for a new partition,
we first try to fetch it in the hash table, if found, we can avoid init
a new one and just re-use the opened one by rescan it, otherwise, we init
a new one and store it in the hash table. The cache is destoryed when
EndPlan.

It seems not easy to add test case for this memory leak issue, the detailed
test cases can be found in the following github link, we have manually test
it. Pipeline should cover the correct of the modification of executor code.

See Github Issue: https://github.com/greenplum-db/gpdb/issues/12533,
this commit fixes it.

Co-authored-by: Junfeng Yang <yjerome@vmware.com>
Co-authored-by: Chris Hajas  <chajas@vmware.com>
Co-authored-by: David Kimura <dkimura@vmware.com>
Reviewed-by:       Shreedhar Hardikar <shardikar@vmware.com>
(cherry picked from commit d837bc3f696dd10150d30c8b0a92826a51c8f1ff)