Removing Recheck and Filter Condition's columns from Bitmap Heap Scan output list when using ORCA. (#284)
The changes introduced in PR #12176 revealed a problem with excessive Bitmap Heap Scan's targetlist which includes Recheck and Filter Condition's columns (introduced in old 97edf9b).
Memory tuple binding generated based on targetlist with different count of attrs may point to different offsets and so, to some garbage or out of range memory area. See issue #12796 for more.
Additional regression test shows how to reproduce a bug before fix. A couple of tests was modified, because there is no need of projection in them started from this patch.
Natively raise exception in case of targetlist mismatch.