Fix query with mode() crash when QueryFinishPending is true (#297)
In case of query with mode() mode_final scans tuples and counts
frequencies in a loop using tuplesort_getdatum to fetch tuples.

In case of MK tuplesort state tuplesort_getdatum calls
tuplesort_getdatum_mk that uses tuplesort_gettuple_common_pos
to fetch the next tuple.

There is a check for QueryFinishPending flag and when it is set
to true return with false (no next tuple).

If QueryFinishPending is true before entry to scan tuples loop
in mode_final then no tuple value is fetched. That leaves last_val
unchanged from init value 0 and attempting to pfree it results in
a crash.

New test verifies the fix.