Extend PgBackendStatus with waiting event start time. (#316)
As we want to somehow monitor main waiting events (PGBE_WAITING_LOCK, PGBE_WAITING_REPLICATION) we need to add a time events initially started.
We also need to clear it once awaiting is over (PGBE_WAITING_NONE). To handle this, we add new st_waiting_start_timestamp field and working with it when track_activities GUC is enabled.