Setup SIGUSR1 procsignal_sigusr1_handler handler.
This is normally automatically done when a bgworker declares the
BGWORKER_BACKEND_DATABASE_CONNECTION flag, but our bgworker doesn't connect to
databases, even though it calls InitPostgres, which will still initialize a new
bacckend and thus participate to the ProcSignal infrastructure.

This wasn't a problem until recently, but now that DROP DATABASE relies on
PROCSIG_BARRIER to work on Windows (see commit
4eb2176318d0561846c1f9fb3c68bede799d640f) we need to respond to ProcSignal
notifications.