[ADBDEV-2501] Wrong subplan order when a query has a partitioned table  (#330)
By the time we get in nested_subplan_mutator, the plan already has dedicated
subplans for each partition. The intention of this part of
adjust_appendrel_attrs_mutator is to create a copy of the
most-closely-nested subplan and add it to subplans list. It's implied that
nested subplans of a copy will be added later in fixup_subplans, but it's the
cause of a crush: more nested subplans will be added after their parent subplans
in a flat list. The intension of this fix is to recursively traverse through
all the subplans and add them right on the spot. We make subplan copies
is nested_subplan_mutator and expect that when we get to fixup_subplans,
all the nested subplans are already initialized and fixup_subplans
doens't need to make any copies anymore.