[ADBDEV-1938] Explain provides wrong rows number when flow is missing (#292)
When we ask GPDB to explain a query, it uses flow to count rows number, and it can be either
a single segment command, a command can be executed on all segments or a limited number
of segments in general cases. But there are cases when the flow is not generated, so GPDB uses
segments count by default, what can be incorrect for some nodes, which are executed on a single
segment only, for example.

This patch doesn't rely on flow, it uses gangSize of a current slice which is calculated correctly
and always presents. gangSize is a simple and reliable way to get a number of segments which
processed some part of a plan. However, orca has another bug related with how it calculates rows 
number. The specialty of orca is that it can return fractional number of rows, and explain should ceil
them later. It can lead to a situation when orca returns a wrong rows number, explain ceils it and we
get an incorrect answer.

Co-authored-by: Dennis Kovalenko <kds@arenadata.io>
Co-authored-by: Vasiliy Ivanov <7vasiliy@gmail.com>