Put some locale related GUCs to sync guc array (#14068)
Note that lc_monetary and lc_time are related to formatting
output data. Besides, formatting functions will be pushed down
to QEs in some common cases. So to keep the output data consistent
with the locale value set in QD, we need to sync these GUCs
between QD and QEs.

Co-authored-by: wuchengwen <wcw190496@alibaba-inc.com>

This is a backport of commit e1a4fdb03edd091b9adc8504c3ed46385241b41c

In the test, the query with the call to 'pg_terminate_backend' has been wrapped
in 'ignore' section. Reason is that on 6x this query is planned in such a way
that the function is called on the segments with arguments (PIDs) collected from
all the cluster. That resulted in several warnings "WARNING: PID XXXXXX is not a
PostgreSQL server process" (and each time the count can be different), before it
actually terminated the backend process.