Disallow the execution of pg_resgroup_get_status and pg_resgroup_get_status_kv functions on entryDB (#390)
The pg_resgroup_get_status and pg_resgroup_get_status_kv functions returned
invalid values when they was called in insert-select queries. The reason for the
error was that the functions was called on entry db only. They should be called
on all segments.
The error can be fixed by setting the proexeclocation property of the functions
to 'i' (PROEXECLOCATION_INITPLAN). But changing catalog for stable release is
something we should avoid unless absolutely necessary for critical issues. So
the execution of pg_resgroup_get_status and pg_resgroup_get_status_kv functions
is disallowed on entryDB. User will get error message and hint how to call the
functions.

https://github.com/greenplum-db/gpdb/issues/14154