gpcheckperf: add buffer size parameter (#428)
Before this patch while running gpcheckperf uitlity the buffer was set by
default for undelying gpnetbenchClient utility as 32Kb. It leaded to problem
with receiving of annoying and misleading warnings about connection between
hosts.

Use '--buffer-size' flag with size in kilobytes to set buffer size,
which will be used at gpnetbenchClient.
It is optional parameter. Default value is 8Kb.