[6X]fix distribution key not exist when creating stage table (#15167)
In order to ensure the integrity of the data, the gpload program first imports the external
data source into the staging table, and then imports it from the staging table into the target table.

Prior to this, [gpload add distribution key for staging table](https://github.com/greenplum-db/gpdb/pull/13159) added a feature to optimize the data
writing performance of the staging table by using the distribution key of the target table as the
distribution key of the staging table to avoid redundant data redistribution.

However, the schema of the staging table and the target table may be not the same, and directly
using the same distribution key may cause the creation of the staging table fail.

An example from a real scenario is as follows:

* The target table contains three fields, namely A, B and C, and the distribution key is A.
* The staging table contains two fields, namely B and C.
* If A is used as the distribution key when creating the staging table, the table creation will fail.

In this PR, firstly check whether the distribution key of the target table is part of the
fields contained in the staging table.

If so, the distribution key of the target table will be directly used as the staging table.

If the result is negative, it will try to use the match_columns field in the configuration
file as the distribution key.

If the field in match_columns is completely contained in the field in the staging table,
use it as the distribution key.

Otherwise no distribution by clause will be added explicitly.
This will result in some slowdown in data import performance, but will ensure a smooth process.
(cherry picked from commit 1c8013a034b8b5623dac2a07e57e02dbaa9a51e9)