Fix "cannot execute squelched plan node of type" error.
This error is caused by the execution logic of Hash Join Node, when the outer
node does not fetch a tuple, it calls the ExecSquelchNode method let the lower
nodes not to produce tuples anymore, leaving the inner node "squalched" but
"un-executed" state. If the hash join needs to be rescanned, then the inner nodes
has to be rescanned too if Hash Node's lefttree does not have chgParam, and finally
a node does not have chgParam but "squelched" will cause ERROR.

This commit resolves the issue by handling of the rescan method in HashJoin Node properly.

Authored-by: wuyuhao28 <wuyuhao28@github.com>
(cherry picked from commit 73517b3121ba9aead136b25ad9a2d0830f673557)