Fix segment segfault when join inner plan contains outer params. (#512)
The call of the cdbpullup_findEclassInTargetList function takes place when
one need to check whether the given EquivalenceClass contains exprs that
can be expressed in terms of the given targetlist. This commit changes the way
how equivalence class is iterated through in order to find a eclass member,
which is already inside the targetList or uses no Vars that are not in
targetList. Previous code returned the first member it found suitable: if the
current member is const or is equal to targetList entry, it will be returned
immediately, otherwise we check whether the member has non-Var type and is not
inside targetList, and return the member or continue the iteration process. That
logic is optimal in most cases, however, there are the cases when that logic
brings erroneous behaviour. For example, in queries containing parameterized
subqueries (like the one presented in the tests) it may lead to passing params
across motion nodes, which may create the situation when these params are being
referenced in different slices, and it probably could lead to segfault when the
executor is accessing its attribute list.

In the query shown in test section such situtation occurs when the Flow node's
hashExprs list is filled during join inner plan creation for scan node. Because
subquery has WHERE clause, the EquivalenceClass for subquery's relation
distribution key contains two members: one is of type T_Param and other
corresponds to distirbution key. When hashExprs list is being filled the
cdbpullup_findEclassInTargetList function is called, and for the given eclass
and targetList it returns the Param member. This Param is returned because
EquivalenceClass member holding it has em_is_const flag on, for situation like
the one in the tests the flag is set anyway, because WHERE operand refers to
outer relations. The function under discussion returns the Param because it
came across the appropriate member without checking targetList. Eventually
this Param entry is appended to hashExprs and later to Seq Scan's targetList
from hashExprs, what leads to invalid nested loop plan and possible segfault
during execution step.

The proposed solution makes cdbpullup_findEclassInTargetList delay the return
of consts and params in order to find a member, which is equal to targetList
entry. This approach solves the issue with param passage, because for such
situations it is guaranteed that plan's targetList contains the distribution
key, and if we iterate through eclass and targetList first, the member
corresponding to distribution key will be returned from the function instead
of the Param. This approach does not break the function's original purpose,
because the entires of eclass may come in arbitrary order.