gpexpand: skip expanded/broken tables when redistributing
This commit is a cherry-pick of 6b7e4c3.
In summary `gpexpand redistribution` earlier used to error out when the table
would have been already expanded manually by the user or the table would have
been damaged. This commit changes this behaviour so that any expanded/broken
tables are skipped without affecting the redistribution for other tables.

Error message was changed due two difference in messages of 6x and 7x.
execSingleton was changed to execSQLForSingleton in mgmt_utils.py to make it
work on 6x. Also, gpexpand.feature test was brought to a common standard by
adding 'rm -rf ~/gpAdminLogs/gpinitsystem*'. Without this line a failing
test would cause other tests to fail.

Co-authored-by: Dmitry Smal <mialinx@yandex-team.ru>
Co-authored-by: Nihal Jain <jnihal@vmware.com>