ADBDEV-3881 Throw error if query has multiple references to modifying CTE (#553)
Multiple references to modifying CTEs can lead to DML operations being
dublicated or cause planner errors. This commit proposes the restriction,
in case of non-fulfillment of which the informative error is thrown. The
if-clause, which checks whether current cte is referenced more than once
and subplan sharing is disabled, is proposed. If it is true, the error is
thrown. This solution should be replaced with more complex logic in future,
allowing materialize and rescan modifying CTE.

Co-Authored-By: Alexey Gordeev <goa@arenadata.io>