Show planner error for SELECT INTO with DML in CTE (#554)
Greenplum fails to execute SELECT INTO and CREATE TABLE AS statements, whose
queries contain modifying CTEs, because Greenplum cannot have two writer
segworker groups for one query. In current implementation all the plans with
modifying CTE contain motion nodes, which separate slice with DML operation
from the slice, where the new table is created. And during execution stage an
error is thrown, telling that node with DML operation cannot be executed,
because one writer segworker already exists. However, showing the error during
planning stage would be more effective, therefore this commit proposes the
display of the error during planning. The if-clause, that checks whether query
contains modifying CTEs, was added. Is is used in block dedicated to SELECT INTO
or CREATE TABLE AS case inside the apply_motion function. If the condition is
true, the proper error is shown.