Add filter to prohibit using encoding for columns at foreign tables (#1139)
Setting ENCODING for columns of foreign tables is useless, but not forbidden
by grammar, so it was silently ignored. This patch adds an error message when
the user tries to explicitly specify the encoding for a foreign table.
We also make sure that GUCs like `default_table_access_method` and
`gp_default_storage_options` are correctly ignored for foreign tables.

The code changes are completely rewritten compared to the original commit.
However, the test is taken from the original commit, with slight
modifications to account for GPDB 7 differences.

(cherry picked from commit bb5a6b85b52161d40aa081dc2c5ec9561eabdb65)