arenadata_toolkit: increase expectation of rows for adb_get_relfilenodes (#608)
It needs for optimizer to prevent hashtable building on the adb_get_relfilenodes
function inside the other joins. It led to high memory usage, because by default
row expectation of functions is 1000. Rows expectation for adb_get_relfilenodes
has been increased to 30000000.

Version of arenadata_toolkit has been increased to 1.1.