Avoid rescan of modifying operations inside correlated Subplans.
When a query had a modifying command inside the correlated SubPlan, the
ModifyTable node could be rescanned for each outer tuple. That lead to
execution errors (rescan of specific nodes is not supported). This happened
because the ParallelizeCorrelatedSubplanMutator function did not expect the
ModifyTable node inside the correlated SubPlans.

This patch adds the support of the ModifyTable nodes for correlated SubPlans.
Currently, ModifyTable node can get into the SubPlan only as the part of CTE
query, therefore, it can either be wrapped in the SubqueryScan node or be
standalone, depending on the SubqueryScan being trivial or not. This
patch affects ParallelizeCorrelatedSubplanMutator function. The patch extends
the if-clause dedicated to the choice of plan nodes, that need to be
broadcasted or focused, and then materialized. The specific conditions related
to modifying operations were added. These conditions checks whether current node
is the SubqueryScan with ModifyTable just under it or current node is a
standalone ModifyTable. If condition is satisfied, node is then processed the
same way as any other Scan-type nodes. Next, the result of ModifyTable is either
broadcasted or focused depending on the target flow type. Then the result is
materialized in order to avoid rescan of underlying nodes.