Fix planning failure on modifying CTE with gp_cte_sharing enabled. (#580)
When the planner was preparing shared scan over CTE, that contained modifying
operation, the Materialize node did not retrieved the target list from
ModifyTable node. That led to the init of Materialize node was failed with
the error: "Material operator does not support projection", because of
that the number of elements in its target list and the child node are different.
It also led to Shared Scan node also doesn't had target list. This happened
because ModifyTable node don't set its own target list on creation.

This patch sets target list of Materialize node on it's creation from returning
lists of child ModifyTable node. We pick first returning list because the first
one (in the case of inherited tables) contains the target list of the root table
while the rest of the child tables. We can take the returning list before it
processing in set_plan_refs because Materialize node has its own processing
in set_plan_refs. Returning list is processed in set_returning_clause_references
which handles situations with outer Vars. Since ModifyTable does the projection,
we don't need to handle these situations.