Fix deadlock in case of the bottleneck HashJoin with shared scan below (#620)
Disabling the prefetch_inner mechanism for HashJoin, when the join is on the
bottleneck, could lead to a deadlock if there is a SharedScan in both parts of
the join. The deadlock occured because SharedScan producer in the inner part of
the join, and consumer in the outer. Thus consumer couldn't perform the shared
scan without prefetching the inner part.

The SharedScan producer for HashJoin is always created under the inner part,
because shareinput_walker function always processes HashJoin node from the Hash
part. For CTEs the prefetching was enabled for all cases, except the bottleneck
join, because CTE path is marked as motionHazard. Therefore, if prefetch_inner
is disabled, the executor can start performing join from the outer part while
the SharedScan producer will be inside the inner part and this will probably
lead to the deadlock if the SharedScan is cross-slice.

This patch fixes this by enabling prefetch_inner for HashJoin on bottleneck when
traversing the plan in shareinput_mutator_dag_to_tree if there are shared scans
in the plan and join is on the bottleneck.