ADBDEV-4349: gplogfilter: fix timerange validation (#627)
Previously the time range filtering of log files was incorrect: there
was a check for belonging a log file to specified time range which
worked like a binary `AND` (if time < begin AND time > end - skip
file), but something like binary `OR` is needed here. Thus file
filtering never worked. Also this validation was executed after the
file was opened, so in case of failed validation there was a redundant
`open/close` operations. This patch fixes such behavior, by:
- moving the logic of belonging log file name the user specified range
  to the separate function (it's also helps to cover corner cases)
- skip file processing, if it doesn't belong to range, at the
  begin of loop to prevent redundant `open/close` operations