Fix CTEs with volatile target list for case of SegmentGeneral and General locuses (#633)
Plan nodes with the General and SegmentGeneral locus type should return the same
result on each of the segments. But a volatile function does not guarantee the
same result for different calls, and therefore if the target list of a plan node
with General or SegmentGeneral locus type contains a volatile function, then the
locus of such a plan node should be brought to SingleQE. In the current
implementation, plan node with volatile function in target list is brought to
SingleQE not for all subqueries. If volatile function is in target list of CTE,
multiset functions and union/except/intersect then they can hide the volatile
function in the underlying nodes and current implementation can't detect them.
The volatile function can be executed on more than one segment and each of them
will work with a different result set which is not correct.

This patch fixes the error by moving the processing of volatile functions in the
target list for plan nodes with the General and SegmentGeneral locus type to
subquery_planner so that processing is performed for all subqueries.