Fix deadlock by shared SegmentGeneral CTE & planning failure by sharing General CTE (#640)
Depending on the usage of the shared CTE with General or SegmentGeneral
subplan, the number of segments the CTE is executed on may be single for joins
with Singleton nodes, and multiple for nodes with other types of locus. In the
current implementation this may lead to deadlock if the CTE is used for both
join targets: the Join with the Singleton node results in the Share Input Scan
producer being executed with 1-gang, while the Join with the N-gang node
creates Share Input Scan reader on multiple segments, so the plan execution
hangs for the SegmentGeneral, and assertions fail inside
shareinput_mutator_xslice_*() functions for the General. If we force execution
of CTE on multiple segments, it will cause redundant motions in case of
joining the CTE with another General or SegmentGeneral. At the moment of
constructing and sharing the CTE we don't know the rest of the plan, so we
can't predict the correct CTE locus. Because replicated tables are considered
small, the most universal and optimal way to fix deadlock would be to inline
CTE scans with General or SegmentGeneral locus.

Inlining plan is implemented by making the CTE subplan once and then copying
it instead of sharing. This solution, however, cannot push filters down, because
this must be done before building the subplan, and the subplan locus can't be
known before it's construction.