Fix deadlock by shared SegmentGeneral CTE & planning failure by sharing General CTE
Depending on the usage of the shared CTE with General or SegmentGeneral
subplan, the number of segments the CTE is executed on may be single for joins
with Singleton nodes, and multiple for nodes with other types of locus. In the
current implementation this may lead to deadlock if the CTE is used for both
join targets: the Join with the Singleton node results in the Share Input Scan
producer being executed with 1-gang, while the Join with the N-gang node
creates Share Input Scan reader on multiple segments, so the plan execution
hangs for the SegmentGeneral, and throw error inside
shareinput_mutator_xslice_2 functions for the General. If we force execution of
CTE on multiple segments, it will cause redundant motions in case of joining the
CTE with another General or SegmentGeneral. At the moment, we do not have the
way to execute a section of the plan on the N segments and take the result from
only one. Because replicated tables are considered small, the most universal and
optimal way to fix deadlock would be to inline CTE scans with General or
SegmentGeneral locus.

Inline is implemented by adding subpath to CtePath and removing subroot from
cteplaninfo. This should be done after processing the volatile functions in the
target list and having qual so as not to do inline in case this part of the plan
will still be executed on singleQE.

(cherry picked from commit 6cac355c7cfc4684018b210697cec37b0d7dd021)

Differences from 6X:
	The inline method is adapted for 7X.
	The test for replicated tables has been changed so that it reproduces the
problem on 7X.

Co-authored-by: Viktor Kurilko <v.kurilko@arenadata.io>