Fix missing query info in gpperfmon when query text differs from the executed statement (#646)
Gpperfmon was tracking active queries and filtering out the ones with finished
or inactive status. The inactive status was detected by missing of a record with
corresponding sess_id and query text in pg_stat_activity. The active query text
is transferred to gpperfmon by executor in .txt file, and in some cases the
query text in executor may differ from the one in pg_stat_activity - for
example, in a case of running a query from DO block: pg_stat_activity contains
the whole DO block text, while the executor writes the query executed. Thus this
query would be filtered out by gpperfmon, removed from queries_now and the .txt
file would be removed. If the file with query text was removed before the
execution end, then gpperfmon added a record in queries_history without query
text.

As all queries are initiated by PortalRun(), it guards the execution with
PG_TRY/PG_CATCH block, which ensures that the finished status will be sent to
gpperfmon, and the query .txt file will be removed in case of both successful
and unsuccessful execution. Thus, the query text comparison is redundant and
leads to losing query text and removal of the query from queries_now.

This patch fixes the loss of records in queries_now and the lack of the query
text in queries_history by removing the comparison of the query text. To detect
backend crash we check that session doesn't exist in pg_stat_activity. This
check is implemented as searching in the active sessions set which replaces
active_query_tab. The check interval of 5 times has been removed since the
search in the hash table is fast enough. The get_active_queries function has
been replaced with get_active_sessions, which creates a set of session ids.

In the test, we execute a query in the DO block for 80 seconds to reproduce the
problem without this patch. It takes 80 seconds because it takes 5 * quantum
before query was checked by is_query_not_active and did not get into the bloom
filter. After 3 * quantum, if the bloom filter check return false, file will be
removed. During the tests quantum is 10 seconds.

Co-authored by v.kurilko@arenadata.io